+++ solver.cpp
@@ -19,7 +19,7 @@ double right_expr_z2(double x, double z2, double z1)
 //// ФУНКЦИИ ЧИСЛЕННЫХ МЕТОДОВ РЕШЕНИЯ СИСТЕМЫ ДИФФЕРЕНЦИАЛЬНЫХ УРАВНЕНИЙ:
 
 // Выполняет шаг метода Рунге-Кутты 4-го порядка.
-void rk_step(uint idx, double *arg, double *cur, double *other, double step, double (*expr)(double, double, double))
+void rk_step(size_t idx, double *arg, double *cur, double *other, double step, double (*expr)(double, double, double))
 {
 	double k1 = expr(arg[idx - 1], cur[idx - 1], other[idx - 1]);
 	double k2 = expr(arg[idx - 1] + step / 2, cur[idx - 1] + k1 * step / 2, other[idx - 1]);
@@ -30,7 +30,7 @@ void rk_step(uint idx, double *arg, double *cur, double *other, double step, dou
 }
 
 // Выполняет шаг метода Адамса-Мултона 5-го порядка.
-void adams_step(uint idx, double *arg, double *cur, double *other, double step, double (*expr)(double, double, double))
+void adams_step(size_t idx, double *arg, double *cur, double *other, double step, double (*expr)(double, double, double))
 {
 	double predictor_cfs[ACC_ORDER] = { 1901.0 / 720, 1387.0 / 360, 109.0 / 30, 637.0 / 360, 251.0 / 720 },
 		   corrector_cfs[ACC_ORDER] = { 251.0 / 720, 646.0 / 720, 264.0 / 720, 106.0 / 720, 19.0 / 720 };
@@ -38,12 +38,12 @@ void adams_step(uint idx, double *arg, double *cur, double *other, double step, 
 	double predictor_value = 0,
 		   corrector_value = 0;
 
-	for (uint i = 0; i < ACC_ORDER; ++i)
+	for (size_t i = 0; i < ACC_ORDER; ++i)
 		predictor_value += predictor_cfs[i] * expr(arg[idx - i - 1], cur[idx - i - 1], other[idx - i - 1]) * (i & 1 ? -1 : 1);
 
 	cur[idx] = cur[idx - 1] + step * predictor_value;
 
-	for (uint i = 0; i < ACC_ORDER; ++i)
+	for (size_t i = 0; i < ACC_ORDER; ++i)
 		corrector_value += corrector_cfs[i] * expr(arg[idx - i], cur[idx - i], other[!i ? idx - 1 : idx - i]) * (!i || i & 1 ? 1 : -1);
 
 	cur[idx] = cur[idx - 1] + step * corrector_value;
@@ -61,13 +61,13 @@ double rk_adams_solve(BoundaryData *data, double der_a, double *res)
 
 	res[0] = data->func_a;
 
-	for (uint i = 1; i < ACC_ORDER; ++i)
+	for (size_t i = 1; i < ACC_ORDER; ++i)
 	{
 		rk_step(i, x_array, res, z_array, step, right_expr_z1);
 		rk_step(i, x_array, z_array, res, step, right_expr_z2);
 	}
 
-	for (uint i = ACC_ORDER; i < data->intervals + 1; ++i)
+	for (size_t i = ACC_ORDER; i < data->intervals + 1; ++i)
 	{
 		adams_step(i, x_array, res, z_array, step, right_expr_z1);
 		adams_step(i, x_array, z_array, res, step, right_expr_z2);
@@ -112,11 +112,11 @@ void newton_solve(BoundaryData *data, double arg, double *res, double eps)
 //// ФУНКЦИИ МЕТОДОВ АПОСТЕРИОРНОЙ ОЦЕНКИ ПОГРЕШНОСТИ ПО ПРАВИЛУ РУНГЕ:
 
 // Возвращает евклидово расстояние между двумя векторами.
-double vector_distance(uint len, double *y1, double *y2)
+double vector_distance(size_t len, double *y1, double *y2)
 {
 	double res = 0;
 
-	for (uint i = 0; i < len; ++i)
+	for (size_t i = 0; i < len; ++i)
 		res += pow(y1[i] - y2[2 * i], 2);
 
 	return sqrt(res) / len;
@@ -156,14 +156,14 @@ void runge_error_solve(BoundaryData *data, double arg, double **res, double eps)
 
 // Заполняет массив значениями с равномерным шагом в заданном диапазоне.
 // Возвращает разность между двумя последовательными значениями в массиве.
-double step_fill(uint len, double *arr, double start, double stop)
+double step_fill(size_t len, double *arr, double start, double stop)
 {
 	double step = (stop - start) / (len - 1);
 
 	arr[0] = start;
 	arr[len - 1] = stop;
 
-	for (uint i = 1; i < len - 1; ++i)
+	for (size_t i = 1; i < len - 1; ++i)
 		arr[i] = arr[i - 1] + step;
 
 	return step;
@@ -171,8 +171,8 @@ double step_fill(uint len, double *arr, double start, double stop)
 
 // Заполняет массив значениями функции на заданных в другом массиве значениях аргумента.
 // Подразумевается, что массивы значений аргумента и функции имеют одинаковую длину.
-void func_fill(uint len, double *arg, double *res, double (*func)(double))
+void func_fill(size_t len, double *arg, double *res, double (*func)(double))
 {
-    for (uint i = 0; i < len; ++i)
+    for (size_t i = 0; i < len; ++i)
         res[i] = func(arg[i]);
 }
